Update opinion.go
Panic was caused by some unknown problems in the system framework, may occurs when calling some particular function. This was fixed by calling some redundant function. The changes are only in this file, by replacing the opinion.go in your thread, the panic issue will be resolved.

Also, Reputation Matrix was fixed. Now, using a sigmoid function, the function gives a evaluation of reputation based on: 
1.1. the energy contribution of other agent 
1.2. the energy contribution of itself 
2.    the compensation of leaving bike